finish individual views for the claims
