+++ app/src/navigation/screens/ClaimScreens/AdultClaimScreen.tsx
@@ -1,24 +1,23 @@
-import { ScrollView, View, StatusBar } from "react-native";
-import ClaimScreenStyles from "../../../styles/ClaimScreenStyles";
-import commonStyles from "../../../styles/styles";
-import { getUserClaimByType } from "../../../utils/claim-utils";
-import { ClaimTypeConstants } from "../../../constants/common";
-import { Input } from "@rneui/themed";
-import React from "react";
-import { FormState } from "../../../types/claim";
-import { Button } from "@rneui/base";
-import useBaseClaim from "../../../hooks/useBaseClaim";
-import { ProfileStackNavigation } from "../../../types/navigation";
 import { useNavigation } from "@react-navigation/native";
-import VerificationFiles from "../../../components/VerificationFiles";
+import React from "react";
+import { ScrollView, StatusBar, View, Text } from "react-native";
+import { ClaimTypeConstants } from "../../../constants/common";
+import useBaseClaim from "../../../hooks/useBaseClaim";
 import useClaimScreen from "../../../hooks/useClaimScreen";
+import ClaimScreenStyles from "../../../styles/ClaimScreenStyles";
+import { FormState } from "../../../types/claim";
+import { ProfileStackNavigation } from "../../../types/navigation";
+import { getUserClaimByType } from "../../../utils/claim-utils";
+import commonStyles from "../../../styles/styles";
+import { Button, Switch } from "@rneui/themed";
+import VerificationFiles from "../../../components/VerificationFiles";
 
-type Navigation = ProfileStackNavigation<"NameClaim">;
+type Navigation = ProfileStackNavigation<"AdultClaim">;
 
-const NameClaimScreen: React.FC = () => {
+const AdultClaimScreen: React.FC = () => {
   const navigation = useNavigation<Navigation>();
   const { claim, userClaim } = getUserClaimByType(
-    ClaimTypeConstants.NameCredential
+    ClaimTypeConstants.AdultCredential
   );
   const [formState, setFormState] = React.useState<FormState>(
     userClaim?.value ?? {}
@@ -33,7 +32,10 @@ const NameClaimScreen: React.FC = () => {
   const { onSelectFile, selectedFileIds, setSelectedFileIds } =
     useClaimScreen();
 
-  const canSave = formState["firstName"] && formState["lastName"];
+  const canSave =
+    claim.fields.filter((field) => formState[field.id]).length ===
+      claim.fields.length &&
+    ((isVerifying && selectedFileIds.length > 0) || !isVerifying);
 
   return (
     <View style={commonStyles.screen}>
@@ -48,13 +50,13 @@ const NameClaimScreen: React.FC = () => {
               }));
             };
             return (
-              <View key={field.id}>
-                <Input
-                  label={field.title}
-                  clearButtonMode="always"
-                  autoCapitalize="none"
-                  value={formState[field.id] as string}
-                  onChangeText={onChange}
+              <View key={field.id} style={{ paddingVertical: 20 }}>
+                <Text style={{ marginBottom: 20 }}>{field.title}</Text>
+                <Switch
+                  value={formState[field.id] === "true"}
+                  onValueChange={(value) => {
+                    onChange(value ? "true" : "false");
+                  }}
                 />
               </View>
             );
@@ -85,4 +87,4 @@ const NameClaimScreen: React.FC = () => {
   );
 };
 
-export default NameClaimScreen;
+export default AdultClaimScreen;
